requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
mysql-connector-python==8.2.0
requests==2.31.0
beautifulsoup4==4.12.2
redis==5.0.1
gunicorn==21.2.0